+++ src/components/Navbar.js
@@ -38,7 +38,7 @@ const NavTitle = () => {
                 paddingLeft: "10px",
               }}
             >
-              Pulsa
+              HolaPulsa
             </h1>
           </a>
           <button
