+++ src/pages/Login.jsx
@@ -1,31 +1,33 @@
 import React, { useState, useEffect } from "react";
 import Swal from "sweetalert2";
+import { useDispatch } from "react-redux";
+import { addTokenAdmin } from "../stores/auth";
 import { useNavigate, Link } from "react-router-dom";
+import "../assets/style/login.css";
 import NavTitle from "../components/NavTitle";
-import "../assets/style/register.css";
 
-export default function Register() {
-  const [usernameR, setUserName] = useState();
-  const [passwordR, setPassword] = useState();
-  const [cofPasswordR, setCofPassword] = useState();
+export default function Login() {
+  const dispatch = useDispatch();
+  const [usernameR, setUserName] = useState("");
+  const [passwordR, setPassword] = useState("");
   const [loading, setLoading] = useState(false);
+  const [tempAdmin, setTempAdmin] = useState({
+    token: "4gfnc1",
+    id: 1,
+    username: "Budi",
+  });
   const navigate = useNavigate();
 
   useEffect(() => {
     async function cek() {
-      if (passwordR != cofPasswordR) {
-        alert("Password yang dimasukkan tidak sesuai dengan yang sebelumnya");
-      } else {
-        Swal.fire({
-          title: "Create New Accout Succes!",
-          text: "Your Account has Already Created, Now You can Login!",
-          type: "success",
-        }).then(function () {
-          console.log("test");
-          navigate("/login");
-        });
-      }
-      setLoading(false);
+      dispatch(addTokenAdmin(tempAdmin));
+      Swal.fire({
+        title: "Login Succes!",
+        text: "Welcome!",
+        icon: "success",
+      }).then(function () {
+        navigate("/");
+      });
     }
     if (loading) {
       cek();
@@ -38,43 +40,35 @@ export default function Register() {
   };
 
   return (
-    <div className="register">
+    <div className="login">
       <NavTitle></NavTitle>
       <div className="space">
-        <h1 className="text-center">Register</h1>
+        <h1 className="text-center">Login</h1>
         <form onSubmit={handleSubmit}>
-          <p>New Username</p>
+          <p>Username</p>
           <div class="input-group mb-3">
             <input
               type="text"
               class="form-control"
-              placeholder="New Username..."
+              placeholder="Username..."
+              name="username"
               onChange={(e) => setUserName(e.target.value)}
               aria-label="Username"
               aria-describedby="basic-addon1"
               required
             />
           </div>
+
           <p>Password</p>
+
           <div class="input-group mb-3">
             <input
               type="password"
               class="form-control"
               placeholder="Password..."
+              name="password"
               onChange={(e) => setPassword(e.target.value)}
               aria-label="password"
-              aria-describedby="basic-addon1"
-              required
-            />
-          </div>
-          <p>Confirm Password</p>
-          <div class="input-group mb-3">
-            <input
-              type="password"
-              class="form-control"
-              placeholder="Confirm Password..."
-              onChange={(e) => setCofPassword(e.target.value)}
-              aria-label="confirmPassword"
               aria-describedby="basic-addon1"
               required
             />
@@ -82,18 +76,18 @@ export default function Register() {
           <div className="tombol">
             <center>
               <button
-                className="btn btn-outline-primary"
                 type="submit"
+                className="btn btn-outline-primary"
                 data-testid="enter"
               >
-                Make New Account
+                Submit
               </button>
             </center>
           </div>
         </form>
         <div className="text-center">
           <p>
-            Already Have Account? <Link to="/login">Login</Link>
+            Dont Have Account? <Link to="/register">Register</Link>
           </p>
         </div>
       </div>
